Reworked build scripts
build-debug.bat and build-release.bat continue to be simple just run to compile, regardless of current working path.

However, they now support specifying a path to the src folder to build.

.obj, .debug, and .release will be created on the path of the build-debug.bat and build-release.bat

---

There is also a bat folder that build-debug.bat and build-release.bat need to work. There are build-debug-actual.bat and build-release-actual.bat.

---

build-debug-actual.bat and build-release-actual.bat require a path given as parameter to the src folder to work. They will create .obj, .debug, and .release on the current working path.